Updated Dockerfile for latest pysaml2 and SATOSA
Updated Dockerfile to use release 4.7.0 for pysaml2 instead of a
a specific pull request (PR) since the PR was merged. Also updated to
use a more current commit for SATOSA. Since the microservice have been
merged back into SATOSA for the time being instead of their own
repository, also changed configuration for plugins to find the modules
in the satosa namespace.